// generated/src/app/api-models/cap-incentive-quality-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapIncentiveQualityWork {
  seqCcalcId: number;
  seqCapIncentQltyWork: number;
  seqCapPoolcalcWork: number;
  seqIncentiveRule: number;
  seqQualityPgm: number;
  seqProvContract: number;
  qualityIncentiveAmount: number;
  qualityIncentivePercent: number;
  qualityCapApply: string;
  incentiveRetro: string;
}